requirements.txt
annotated-types==0.7.0
anyio==4.10.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.116.1
google-api-core==2.25.1
google-api-python-client==2.181.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-storage==3.3.1
google-crc32c==1.7.1
google-oauth2-tool==0.0.3
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httplib2==0.30.0
Hypercorn==0.17.3
hyperframe==6.1.0
idna==3.10
jose==1.0.0
oauth2client==4.1.3
priority==2.0.0
proto-plus==1.26.1
protobuf==6.32.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
pymongo==4.14.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
requests==2.32.5
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.47.3
typing-inspection==0.4.1
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
wsproto==1.2.0
razorpay==1.4.2
